+++ config.yaml
@@ -1,2 +1,2 @@
 theme: emacs
-message_box: ASCII
+box_style: ASCII
